Move LadspaEditor to new file
